Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,7 +1,7 @@
-#include <stdio.h>
+#include<stdio.h>
 int main()
 {
-    int m;
+    float m;
     scanf("%f",&m);
     printf("%f",m);
     return 0;
